% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objectives.R
\name{RMSE}
\alias{RMSE}
\title{Root mean squared error}
\usage{
RMSE(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{Observed data series}

\item{y}{Estimated/modelled values}

\item{na.rm}{NA remove defaulted as FALSE}
}
\value{
The Root Mean Square Error value
}
\description{
Calculate RMSE model diagnostic from an observed series,
x, and modelled series, y.
}
\examples{
x <- c(3, 3, 4, 20, 62, 85, 33, 20, 5)
y <- c(5, 12, 65, 43, 34, 28, 13, 12, 22)
RMSE(x, y)
}
